handlers: split each handler to its own file
