+++ src/libcrun/custom-handler.h
@@ -23,10 +23,10 @@
 #include <stdio.h>
 
 enum handler_configure_phase
-  {
-    HANDLER_CONFIGURE_BEFORE_MOUNTS = 1,
-    HANDLER_CONFIGURE_AFTER_MOUNTS,
-  };
+{
+  HANDLER_CONFIGURE_BEFORE_MOUNTS = 1,
+  HANDLER_CONFIGURE_AFTER_MOUNTS,
+};
 
 struct custom_handler_s
 {
